Refactor: [BADA-68] User enum 패키지 분리 #29
