+++ src/main/java/com/TwoSeaU/BaData/domain/auth/service/SocialUserService.java
@@ -4,17 +4,14 @@ import com.TwoSeaU.BaData.domain.auth.jwt.ServiceTokenProvider;
 import com.TwoSeaU.BaData.domain.auth.dto.oauth2.request.GetOAuth2UserProfileRequest;
 import com.TwoSeaU.BaData.domain.auth.dto.response.IssueTokenUserStatusResponse;
 import com.TwoSeaU.BaData.domain.auth.dto.response.IssueServiceTokenResponse;
-import com.TwoSeaU.BaData.domain.user.entity.Role;
-import com.TwoSeaU.BaData.domain.user.entity.SocialType;
+import com.TwoSeaU.BaData.domain.user.enums.Role;
+import com.TwoSeaU.BaData.domain.user.enums.SocialType;
 import com.TwoSeaU.BaData.domain.user.entity.User;
-import com.TwoSeaU.BaData.domain.user.exception.UserException;
 import com.TwoSeaU.BaData.domain.user.repository.UserRepository;
 import com.TwoSeaU.BaData.global.redis.RedisUtil;
-import com.TwoSeaU.BaData.global.response.GeneralException;
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
