--- src/main/java/com/TwoSeaU/BaData/domain/user/entity/Role.java
@@ -1,5 +1,0 @@
-package com.TwoSeaU.BaData.domain.user.entity;
-
-public enum Role {
-    ADMIN,GENERAL
-}
