+++ src/main/java/com/TwoSeaU/BaData/domain/user/enums/Role.java
@@ -1,0 +1,5 @@
+package com.TwoSeaU.BaData.domain.user.enums;
+
+public enum Role {
+    ADMIN,GENERAL
+}
